pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "promptman"
version = "0.3.2"
description = "AI Prompt Management System"
authors = ["MakerCorn <corn.mendoza@outlook.com>"]
license = "Non-Commercial"
readme = "README.md"
packages = [
    { include = "prompt_manager.py" }, 
    { include = "prompt_data_manager.py" },
    { include = "auth_manager.py" },
    { include = "api_token_manager.py" },
    { include = "api_endpoints.py" },
    { include = "langwatch_optimizer.py" },
    { include = "run.py" },
    { include = "token_calculator.py" },
    { include = "__main__.py" },
    { include = "src" }
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "e2e: End-to-end tests",
    "slow: Tests that take a long time to run",
]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.poetry.dependencies]
python = "^3.12"

# Core UI Framework
gradio = ">=5.32.1,<6.0.0"

# HTTP Requests and API
requests = ">=2.32.3,<3.0.0"
fastapi = ">=0.115.6,<1.0.0"
pydantic = ">=2.10.3,<3.0.0"

# Database
psycopg2-binary = ">=2.9.10,<3.0.0"

# Configuration and Environment
python-dotenv = ">=1.1.0,<2.0.0"

# Authentication and Security
pyjwt = ">=2.10.1,<3.0.0"
cryptography = ">=43.0.3,<44.0.0"

# AI and Tokenization (optional)
tiktoken = ">=0.8.0,<1.0.0"

# Optional password hashing libraries
bcrypt = {version = ">=4.0.0,<5.0.0", optional = true}
argon2-cffi = {version = ">=21.0.0,<24.0.0", optional = true}

[tool.poetry.extras]
password-hashing = ["bcrypt", "argon2-cffi"]

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = ">=0.34.0,<1.0.0"}
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"

# Code quality tools (optional)
flake8 = {version = "^6.0.0", optional = true}
mypy = {version = "^1.0.0", optional = true}
black = {version = "^23.0.0", optional = true}
isort = "^6.0.1"
bandit = "^1.8.5"

[tool.poetry.group.test.dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"

[tool.poetry.group.e2e.dependencies]
selenium = "^4.15.0"
playwright = "^1.40.0"
requests = "^2.31.0"
pytest-xdist = "^3.3.0"
pytest-html = "^4.1.0"

# Black configuration
[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort configuration (compatible with black)
[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true